добавлены тесты

+++ src/kv.c
@@ -1,6 +1,6 @@
 #include "kv.h"
 
-double kv(double a,  double b,  double c, double *i, double *o)
+int kv(double a,  double b,  double c, double *i, double *o)
 {
 	if ((a == 0) || ((a == 0) && (b == 0)) || ((a == 0) && (b == 0) && (c == 0)))
 		return INVALUD_ARGUMENTS;
